scripts: dedup GitHub API helper + block claude_*.py in gitignore
- Extract the duplicated `gh_api()` + REPO constant from metrics.py and
  check_traction.py into a shared scripts/_gh_api.py (both scripts queried the
  same repo stats/traffic/releases endpoints via divergent copies). Both now
  import from it. scripts/test-metrics.py passes (22 tests).
- .gitignore: add scripts/claude_*.py — the ignore list only had
  scripts/claude-*.sh (bash), so the harness's Python hook ports (claude_
  format.py etc.) could leak into this public repo. No tracked claude_*.py
  exists, so nothing legitimate is hidden. Closes the same gap PR #50 addressed
  for the .sh case.
